+++ src/Symfony/Component/DependencyInjection/Compiler/ResolveParameterPlaceHoldersPass.php
@@ -23,6 +23,12 @@ use Symfony\Component\DependencyInjection\Exception\ParameterNotFoundException;
 class ResolveParameterPlaceHoldersPass extends AbstractRecursivePass
 {
     private $bag;
+    private $resolveArrays;
+
+    public function __construct($resolveArrays = true)
+    {
+        $this->resolveArrays = $resolveArrays;
+    }
 
     /**
      * {@inheritdoc}
@@ -55,7 +61,9 @@ class ResolveParameterPlaceHoldersPass extends AbstractRecursivePass
     protected function processValue($value, $isRoot = false)
     {
         if (is_string($value)) {
-            return $this->bag->resolveValue($value);
+            $v = $this->bag->resolveValue($value);
+
+            return $this->resolveArrays || !$v || !is_array($v) ? $v : $value;
         }
         if ($value instanceof Definition) {
             $changes = $value->getChanges();
@@ -65,10 +73,14 @@ class ResolveParameterPlaceHoldersPass extends AbstractRecursivePass
             if (isset($changes['file'])) {
                 $value->setFile($this->bag->resolveValue($value->getFile()));
             }
-            $value->setProperties($this->bag->resolveValue($value->getProperties()));
-            $value->setMethodCalls($this->bag->resolveValue($value->getMethodCalls()));
         }
 
-        return parent::processValue($value, $isRoot);
+        $value = parent::processValue($value, $isRoot);
+
+        if ($value && is_array($value)) {
+            $value = array_combine($this->bag->resolveValue(array_keys($value)), $value);
+        }
+
+        return $value;
     }
 }