+++ src/Symfony/Bundle/SecurityBundle/CHANGELOG.md
@@ -8,6 +8,15 @@ CHANGELOG
  * made `FirewallMap::$container` and `::$map` private
  * made the first `UserPasswordEncoderCommand::_construct()` argument mandatory
  * `UserPasswordEncoderCommand` does not extend `ContainerAwareCommand` anymore
+ * removed support for voters that don't implement the `VoterInterface`
+
+3.4.0
+-----
+
+ * Tagging voters with the `security.voter` tag without implementing the
+   `VoterInterface` on the class is now deprecated and will be removed in 4.0.
+ * [BC BREAK] `FirewallContext::getListeners()` now returns `\Traversable|array`
+ * added info about called security listeners in profiler 
 
 3.3.0
 -----